django_test/core/api/module/Detection/requirements.txt
absl-py==2.3.1
ai-edge-litert==2.0.2
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backports.strenum==1.2.8
beautifulsoup4==4.13.5
bleach==6.2.0
bs4==0.0.2
cattrs==24.1.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
docstring-to-markdown==0.16
exceptiongroup==1.2.2
executing==2.2.1
fake-useragent==2.2.0
fastjsonschema==2.21.2
filelock==3.13.1
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.0
fqdn==1.5.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.7
importlib_metadata==8.6.1
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jedi-language-server==0.45.0
Jinja2==3.1.4
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.11.3
kiwisolver==1.4.9
lark==1.3.0
libclang==18.1.1
lsprotocol==2023.0.1
lxml==6.0.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
microvenv==2025.0
mistune==3.1.4
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.1.3
onnx==1.19.0
onnx2tf==1.28.2
onnx_graphsurgeon==0.5.8
onnxruntime==1.22.1
onnxslim==0.1.69
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pexpect==4.9.0
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs==4.4.0
pluggy==1.6.0
polars==1.33.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==5.29.5
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pyee==11.1.1
pygls==1.3.1
Pygments==2.19.2
pyparsing==3.2.5
pyppeteer==2.0.0
pyquery==2.0.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
requests-html==0.10.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
roboflow==1.2.9
rpds-py==0.27.1
scipy==1.16.2
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sng4onnx==1.0.4
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
tf_keras==2.19.0
tinycss2==1.4.0
tomli==2.2.1
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.13.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
uri-template==1.3.0
urllib3==1.26.20
w3lib==2.3.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
zipp==3.21.0
